+++ docs/examples/docs-demo-six-date-10.vue
@@ -1,0 +1,16 @@
+<template>
+<div>
+
+        <six-date label="Date with d.M.yyyy format" date-format="d.M.yyyy"></six-date>
+      
+</div>
+</template>
+<style>
+
+</style>
+<script>
+export default {
+  name: 'docs-demo-six-date-10',
+  mounted() {  }
+}
+</script>